% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_Nozzle_Characteristics.R
\name{charact_cal}
\alias{charact_cal}
\title{Calculate three velocity components}
\usage{
charact_cal(app_p, angle, rhosoln)
}
\arguments{
\item{app_p}{Nozzle Pressure}

\item{angle}{Nozzle Angle}

\item{rhosoln}{Mix Density}
}
\value{

}
\description{
Calculate three velocity components
}
