% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_Deposition_Calcs_function.R
\name{deposition_calcs}
\alias{deposition_calcs}
\title{Calculate Deposition}
\usage{
deposition_calcs(
  IAR,
  xactive,
  FD,
  PL,
  NozzleSpacing,
  psipsipsi,
  rhoL,
  Cent_inp,
  Dwnd_inp,
  Uwnd_inp,
  Dpmax,
  DDpmin,
  a,
  MMM,
  lambda,
  Driver
)
}
\arguments{
\item{IAR}{teststes}

\item{Driver}{"text","Shiny", "Silent"}
}
\value{

}
\description{
Calculate Deposition
}
