% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_Droplet_Transport_function.R
\name{droplet_transport}
\alias{droplet_transport}
\title{Calculating droplet transportation}
\usage{
droplet_transport(
  Tair,
  RH,
  rhow,
  rhos,
  xs0,
  H0,
  DTwb,
  hcm,
  Uf,
  z0,
  Pn,
  vz0,
  vx0,
  ddd_inp,
  Driver
)
}
\arguments{
\item{Tair}{}

\item{RH}{}

\item{rhow}{}

\item{rhos}{}

\item{xs0}{}

\item{H0}{}

\item{DTwb}{}

\item{hcm}{}

\item{Uf}{}

\item{z0}{}

\item{Pn}{}

\item{vz0}{}

\item{vx0}{}

\item{ddd_inp}{}

\item{Driver}{}
}
\value{

}
\description{
Calculating droplet transportation
}
