% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs_from_csv.R
\name{inputs_from_csv}
\alias{inputs_from_csv}
\title{Read inputs from the csv files and converts units as needed}
\usage{
inputs_from_csv(DSDData, paramsData, DDDparamsData, paramsID, paramsType)
}
\arguments{
\item{DSDData}{the raw DSDData read from .csv file}

\item{paramsData}{the raw paramsData read from .csv file}

\item{DDDparamsData}{the raw DDDparamsData read from .csv file}

\item{paramsID}{the ID for the parameters}

\item{paramsType}{the type of the parameters}
}
\value{
a list containing all input data
}
\description{
Read inputs from the csv files and converts units as needed
}
