% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_droplet_data.R
\name{plot_droplet_data}
\alias{plot_droplet_data}
\title{Droplet Transportation Calculator}
\usage{
plot_droplet_data(All_droplet_data)
}
\arguments{
\item{All_droplet_data}{Droplet data}
}
\value{
ggplot of droplet data
}
\description{
Droplet Transportation Calculator
}
