% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_psd_function.R
\name{psd}
\alias{psd}
\title{PSD fitting function}
\usage{
psd(y, Dpdata)
}
\arguments{
\item{y}{Average DSD fit data:}

\item{Dpdata}{Corresponding droplet size (in microns):}
}
\value{
list of res=c(a1,a2,d1,d2,k1), plot, and input data.
}
\description{
PSD fitting function
}
