% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_psd_stats.R
\name{psd_stats}
\alias{psd_stats}
\title{PSD statistics}
\usage{
psd_stats(y, Dpdata)
}
\arguments{
\item{y}{Average DSD fit data:}

\item{Dpdata}{Corresponding droplet size (in microns)}
}
\value{
list of statical parameters
}
\description{
PSD statistics
}
