% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psipsipsi_calcs.R
\name{psipsipsi_calcs}
\alias{psipsipsi_calcs}
\title{Calculate Deposition}
\usage{
psipsipsi_calcs()
}
\arguments{
\item{IAR}{Intended Application Rate for Dicamba, lb/acre}
}
\value{

}
\description{
Calculate Deposition
}
