% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCasanova.R
\name{runCasanova}
\alias{runCasanova}
\title{Run Casanova}
\usage{
runCasanova(
  scnFile = "./sample_data/Scenarios.csv",
  DDDparamsFile = "./sample_data/DDD_Params.csv",
  report = F
)
}
\arguments{
\item{scnFile}{a csv file containing the DSD, Params, Param type and param ID used for the analysis; default file is Scenarios.csv}

\item{report}{is a T/F input indicating whether reports need to be printed out}

\item{DSDFile}{a csv file defining the DSD measurements; the function takes the average of each trial}

\item{paramsFile}{a csv file defining all the inputs; default file is Params (Metric) for metric}

\item{DDDParamsFile}{DDD parameter file; default file is DDD_Params.csv}
}
\value{
a list containing all droplet data and deposition for each scenario analyzed
}
\description{
Run Casanova
}
