% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCasanovaApp.R
\name{runCasanovaApp}
\alias{runCasanovaApp}
\title{The function to run the Casanova App}
\usage{
runCasanovaApp()
}
\description{
The function to run the Casanova App
}
\details{
https://deanattali.com/2015/04/21/r-package-shiny-app/ Also, to use private or company repositiories,
https://docs.rstudio.com/connect/1.5.4/admin/package-management.html
}
