% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_wet_bulb_function.R
\name{wet_bulb}
\alias{wet_bulb}
\title{Wet Bulb Depression}
\usage{
wet_bulb(Tair, Patm, RH)
}
\arguments{
\item{Tair}{Dry air temperature (Celcius)}

\item{Patm}{Barometric pressure (mmHg abs)}

\item{RH}{Relative humidity (%}
}
\value{
Change in Wet Bulb Temperature (DTwb in °C) and Wet Bulb Temperature (Twb in °C)
}
\description{
Wet Bulb Depression
}
