% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_report.R
\name{write_report}
\alias{write_report}
\title{Create html report for each scenario}
\usage{
write_report(i, all_inputs, results, report_folder)
}
\arguments{
\item{i}{Scenario}

\item{all_inputs}{list with all model inputs}

\item{results}{the results of the casanova model}
}
\value{
a list containing all input data
}
\description{
Create html report for each scenario
}
