% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_wvprofile_params.R
\name{WV2m}
\alias{WV2m}
\title{Function when 2 wind vs. height measurements are available}
\usage{
WV2m(z1, z2, ux1, ux2)
}
\arguments{
\item{ux2}{}
}
\value{

}
\description{
Function when 2 wind vs. height measurements are available
}
