% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_wvprofile_params.R
\name{wvprofile}
\alias{wvprofile}
\title{Function when 1 wind vs. height measurements are available}
\usage{
wvprofile(z1, ux1, ch)
}
\arguments{
\item{z1}{1st Elevation wind speed (ft):}

\item{ux1}{1st wind speed (mph):}

\item{ch}{Canopy_height (in):}
}
\value{
Friction height, cm (z0) and Friction velocity, cm/sec (uf)
}
\description{
Function when 1 wind vs. height measurements are available
}
