% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_wvprofile_params.R
\name{wvprofilem}
\alias{wvprofilem}
\title{Function when more than 1 wind vs. height measurements are available}
\usage{
wvprofilem(paramsWT, method, ch)
}
\arguments{
\item{paramsWT}{wind speed, temperature vs. elevation}

\item{ch}{crop height}
}
\value{

}
\description{
Function when more than 1 wind vs. height measurements are available
}
