% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_Deposition_Calcs_function.R
\name{deposition_calcs}
\alias{deposition_calcs}
\title{Calculate Deposition}
\usage{
deposition_calcs(
  IAR,
  xactive,
  FD,
  PL,
  NozzleSpacing,
  psipsipsi,
  rhoL,
  Cent_inp,
  Dwnd_inp,
  Uwnd_inp,
  Dpmax,
  DDpmin,
  a,
  MMM,
  lambda,
  Driver
)
}
\arguments{
\item{IAR}{Intended Application Rate for Dicamba, lb/acre}

\item{xactive}{Dicamba conc in tank solution, wtfraction}

\item{FD}{Downwind field depth, ft}

\item{PL}{Crosswind field width, ft}

\item{NozzleSpacing}{Space between nozzles on Boom, inches}

\item{psipsipsi}{Horizontal variation in wind direction around mean direction, 1 stdev, in degrees.}

\item{rhoL}{Density of sprayed solution, grams/cc}

\item{Cent_inp}{output from droplet transportation}

\item{Dwnd_inp}{output from droplet transportation}

\item{Uwnd_inp}{output from droplet transportation}

\item{Dpmax}{D pmax}

\item{DDpmin}{DD pmin}

\item{a}{output from psd function, Calibration results}

\item{MMM}{original value for integration}

\item{lambda}{Controls resolution of deposition calculations; higher numbers increase accuracy}

\item{Driver}{"text","shiny", "Silent"}
}
\value{

}
\description{
Calculate Deposition
}
