% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_Droplet_Transport_function - Copy.R,
%   R/4_Droplet_Transport_function.R
\name{droplet_transport}
\alias{droplet_transport}
\title{Droplet Transportation Calculator}
\usage{
droplet_transport(
  Tair,
  RH,
  rhow,
  rhos,
  xs0,
  H0,
  DTwb,
  hcm,
  Uf,
  z0,
  Pn,
  vz0,
  vx0,
  ddd_inp,
  Driver
)

droplet_transport(
  Tair,
  RH,
  rhow,
  rhos,
  xs0,
  H0,
  DTwb,
  hcm,
  Uf,
  z0,
  Pn,
  vz0,
  vx0,
  ddd_inp,
  Driver
)
}
\arguments{
\item{Tair}{Dry air temperature, degrees C}

\item{RH}{Percent relative humidity}

\item{rhow}{Density of pure water in droplet}

\item{rhos}{Density of dissolved solids in droplet, g/cc}

\item{xs0}{mass fraction total dissolved solids in solution}

\item{H0}{Height of nozzle above ground , inches}

\item{DTwb}{Wetbulb temperature depression, C}

\item{hcm}{Canopy height in cms}

\item{Uf}{Friction velocity, cm/sec (uf)}

\item{z0}{Friction height, cm (z0)}

\item{Pn}{# Nozzle pressure, psi}

\item{vz0}{nozzle characteristics 1}

\item{vx0}{nozzle characteristics 1}

\item{ddd_inp}{ddd parameters}

\item{Driver}{"text","shiny", "Silent"}
}
\value{
Xdist and corresponding transportation.

Xdist and corresponding transportation.
}
\description{
Droplet Transportation Calculator

Droplet Transportation Calculator
}
