% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCasanova.R
\name{runCasanova}
\alias{runCasanova}
\title{Run Casanova}
\usage{
runCasanova(
  scnFile = "./sample_data/Scenarios.csv",
  DDDparamsFile = "./sample_data/DDD_Params.csv",
  report_folder = "./sample_data/reports",
  curve_fit_ini_file = "./sample_data/Curve_Fit_Initial_Values.csv",
  report = T,
  curvefitDSD = F
)
}
\arguments{
\item{scnFile}{a csv file containing the DSD, Params, Param type and param ID used for the analysis; default file is Scenarios.csv}

\item{report_folder}{is the folder to save the .html reports}

\item{report}{is a T/F input indicating whether reports need to be printed out}

\item{curvefitDSD}{is a T/F indicating whether the DSD will be curve fitted or interpolated}

\item{DDDParamsFile}{DDD parameter file; default file is DDD_Params.csv}
}
\value{
a list containing all droplet data and deposition for each scenario analyzed
}
\description{
Run Casanova
}
