% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCasanova.R
\name{runCasanova}
\alias{runCasanova}
\title{Title Run casanova as all}
\usage{
runCasanova(
  dsdTable,
  dsdFile = NULL,
  inputs = "DefaultInputs.R",
  NozzleParamFile = "~/mondep/Casanova/data-raw/Nozzle_Params.csv",
  DDDParamsFile = "~/mondep/Casanova/data-raw/DDD_Params.csv"
)
}
\arguments{
\item{dsdTable}{dsd data as a table, use this when dsdFile is not provided.}

\item{dsdFile}{read in a csv file input and use that as the table}

\item{inputs}{an R file defining all the inputs, and it could be modified!}

\item{NozzleParamFile}{nozzle parameter file}

\item{DDDParamsFile}{DDD parameter file}
}
\value{
a list containing all droplet data and deposition
}
\description{
Title Run casanova as all
}
