% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCasanova.R
\name{runCasanova}
\alias{runCasanova}
\title{Title Run casanova as all}
\usage{
runCasanova(
  dsdTable,
  dsdFile = NULL,
  inputs = "DefaultInputs.R",
  NozzleParamFile = "~/mondep/Casanova/data-raw/Nozzle_Params.csv",
  DDDParamsFile = "~/mondep/Casanova/data-raw/DDD_Params.csv"
)
}
\arguments{
\item{DDDParamsFile}{}
}
\value{

}
\description{
Title Run casanova as all
}
