% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_wvprofile_params.R
\name{WV2m}
\alias{WV2m}
\title{Function when 2 wind vs. height measurements are available}
\usage{
WV2m(z1, z2, ux1, ux2)
}
\arguments{
\item{z1}{elevation of wind velocity in ft}

\item{z2}{elevation of wind velocity in ft}

\item{ux1}{mph wind velocity at elevation}

\item{ux2}{mph wind velocity at elevation}
}
\value{

}
\description{
Function when 2 wind vs. height measurements are available
}
